Running but gives infeasibility
Costruction of lexicographic matrix works well, but cannot build the pareto front. all costrained solution but the first one are infeasible. Bug in the algorythm writing or structural problem? 
It's possible that: 
(a) given that varcost and emissions are not really in competition, that may lead to a bag. To be tried with LCOE
(b) worst: cannot solve with a fixed value of emission because of too little flexibility in the dispatch optimization